DESIGN Notices detail 공지사항 상세 페이지 추가 [병진]

+++ notices/views.py
@@ -17,23 +17,23 @@ def index(request):
 
 # 슈퍼 유저일 경우에만
 def create(request):
-    # if request.user.is_superuser:
-    if request.method == "POST":
-        form = NoticeForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect("notices:index")
+    if request.user.is_superuser:
+        if request.method == "POST":
+            form = NoticeForm(request.POST)
+            if form.is_valid():
+                form.save()
+                return redirect("notices:index")
+        else:
+            form = NoticeForm()
+        return render(
+            request,
+            "notices/create.html",
+            {
+                "form": form,
+            },
+        )
     else:
-        form = NoticeForm()
-    return render(
-        request,
-        "notices/create.html",
-        {
-            "form": form,
-        },
-    )
-    # else:
-    #     return HttpResponseForbidden()
+        return HttpResponseForbidden()
 
 
 def detail(request, pk):
@@ -50,31 +50,31 @@ def detail(request, pk):
 # 슈퍼 유저일 경우에만
 def update(request, pk):
     notice = Notice.objects.get(pk=pk)
-    # if request.user.is_superuser:
-    if request.method == "POST":
-        form = NoticeForm(request.POST, instance=notice)
-        if form.is_valid():
-            form.save()
-            return redirect("notices:detail", pk)
+    if request.user.is_superuser:
+        if request.method == "POST":
+            form = NoticeForm(request.POST, instance=notice)
+            if form.is_valid():
+                form.save()
+                return redirect("notices:detail", pk)
+        else:
+            form = NoticeForm(instance=notice)
+        return render(
+            request,
+            "notices/update.html",
+            {
+                "form": form,
+                "notice": notice,
+            },
+        )
     else:
-        form = NoticeForm(instance=notice)
-    return render(
-        request,
-        "notices/update.html",
-        {
-            "form": form,
-            "notice": notice,
-        },
-    )
-    # else:
-    #     return HttpResponseForbidden()
+        return HttpResponseForbidden()
 
 
 # 슈퍼 유저일 경우에만
 def delete(request, pk):
     notice = Notice.objects.get(pk=pk)
-    # if request.user.is_superuser:
-    notice.delete()
-    return redirect("notices:index")
-    # else:
-    #     return HttpResponseForbidden()
+    if request.user.is_superuser:
+        notice.delete()
+        return redirect("notices:index")
+    else:
+        return HttpResponseForbidden()
